Final email validation in BWC modules is little more than a sanity check
Email validation via regex is almost always incorrect. However, we want to catch as many errors as possible on the client since the server will swallow errors in the context of BWC. To validate the structure of an email address, we test for known invalid characters in the local part and permit almost anything for the domain. TLD's aren't required in the domain part, so really any character `a-z0-9-.` are permitted. However, the first and last characters cannot be `-` or `.`. And an email address can only contain one `@`.

 Since it is unlikely that someone will enter a blatantly bad domain, it's not worthwhile to do anything more than verify that there is only one `@`. Especially since this javascript will sunset once there are no more BWC modules. The bug jira:MAR-2805 was reported internally, so it doesn't seem likely that customers are running into domain errors much.

Fixes jira:MAR-2805.