jira: MAR-2836 Use the correct module for the context when fetching activities
After creating a new attachment activity, the `DragdropAttachmentPlugin` reloads the collection of activities. In some cases, the context has `collectionOptions` object that customizes how the data should be retrieved. For example, `View.Layouts.Base.ActivitystreamLayout` specifies an alternative method for retrieving the activities based on the correct module. When this property exists on the context, `DragdropAttachmentPlugin` should use it to guarantee that the correct activities are requested. When this property does not exist on the context, the default of no special options will be passed when reloading data.