SFA-3548: Fix random failing test
- Hardcode `now()` to always be a specific date
- test to make sure that the diff is what it should be

The issue here was not with the actual code in the app, but how the
expected was setup.  Sometimes it would not generated the correct
start or end date.  It would always be off by one day when it would
fail.

Setting the date to be a static make sense since than can be ran
across multiple timezone at different times of the day.