SC-4487 : Improvements of render in actiondropdown and actionmenu fields.
Children fields are not rendered anymore on `_render` for actiondropdown and actionmenu fields. They're initialized, but they are now rendered on click.
Before, we were trying to render them on '_render', but they had no placeholders so they didn't render anywhere. Then we would actually render them on click on the dropdown button.