Temporary fix for /me hashes (for ETags) (fixes SC-4596)
`/me` depends on data from Metadata. When adding a module, the `modules_list`
of the user changes.
That means that theoretically the user preferences have changed (because the
list of modules have changed).

If we change the `date_modified` of all users, while correct from the
`modules_list` perspective, it isn't true that the user changed his
preferences, instead the system has changed.

The correct fix should be converting `modules_list` into a tuple that contains
the information of the tabs defined by the user (show/hidden and order).
The list of modules that aren't in that tuple, but exist in the `modules_info`
of `/metadata` with `display_tab` set to `true` should be added to the end of
the list of modules in the nav bar.
This will guarantee that `/me` hash is based on user + user preferences and not
based on `/metadata`.