Use the module field when rendering module names
The prior implementation was changing the value of the _module attribute to the formatted module name. The PreviewButtonField attempted to get the metadata for the Targets module during initialization and use the metadata as an object. Since the Targets module is actually known as Prospects in code, the metadata could not be found and there were no defensive checks. It is dangerous to change the _module attribute directly. Instead, ModuleField will format the value when the field is rendered.