Copy should check for attribute existence in the model (fixes SC-5110)
There are cases (preview) where the first render time of the copy field might
not have all the data from the model yet (still fetching the primary address).
When that happens, the field cannot assume that it is in sync. It should wait
until the model provides more information about the fields to keep in sync.