BR-3704: Temporary fix for parity functionality until BR-3708 is resolved
There exists a restricted list of field types which are supported by studio. However this list is not honored when cloning an existing field. For example the email type is not supported through studio, however the clone functionality is still available through studio. Once this issue is resolved we can safely use $field_types as choice validator.