BWC views can only fully unbind on app:view:load (fixes SC-5184)
Since router changes do not mean that the view is disposing, we no longer
provide the `unbindDom` during route changes.

It is removed completely instead of listening to `app:view:load`, because
the previous layout in the controller will be disposed, and during that
dispose it is when we call the `unbindDom()` method.