BR-3743: Use secure require and validate input to it.
VC 7.6 flaw id 239

Note that the require_once shouldn't be necessary as the autoloader should take care of this. However because this is legacy code there might be some corner cases around it. Therefor this fix just addresses the required input validation and secure include without changing the rest of the code flow.