Use notes.email_id and notes.email_type instead of notes.parent_id and notes.parent_type when saving attachments (MAR-3404)
These fields are more or less immutable and allow for a Notes record to have mutable parent fields without the risk of a user accidently breaking the link between an email and its attachment. The notes.email_id field contains the ID of the email or email template to which the note is attached. The notes.email_type field contains either Emails or EmailTemplates, according to the type of email to which the note is attached.