[BR-3895] phpUnit Bug62026Test.testCalculatedFieldCount failure on all stacks
The TeamSetModule bean gets saved before the bean record is inserted/updated into database. It triggers resave of the related beans which in turn triggers recalculation of calculated fields. Since after BR-3366 is merged, the linked beans are no more cached in the link object, the bean being saved is unavailable to the expression, and the result of the calculation is incorrect.

The solution is to save the TeamSetModule record manually in order to avoid premature resave of related beans.