The format of the Message-ID needs to be <unix_timestamp.$id@hostname>
SugarCRM cleans values that appear as HTML in certain cases before inserting that data into the database. Because the Message-ID begins with a "<", an $id that begins with an alphabetic character will cause the Message-ID to be parsed as an invalid HTML tag by HTMLPurifier. Thus $email->message_id is set to an empty string in such a case, causing the Message-ID to not be stored. Prefixing the unix timestamp guarantees that the Message-ID will begin with "<", followed by an integer, which HTMLPurifier will correctly ignore as a non-threatening tag. This allows the Message-ID to be saved to the database whenever appropriate, without risk of losing the value. And it doesn't hurt that it increases the likelihood that each Message-ID is unique, as well.