BR-3535: Disabling the order stability capability for DB2 manager.
This capability was introduced because adding an additional ORDER BY column
caused a lot of performance havoc on list views. After retesting on DB2
v10.5.5 it seems this performance impact is no longer detectable and the
execution plans stay the same with or without the additional ORDER BY
column. See Jira ticket for more details and attached Data Studio
execution plans.