SFA-4195: Fix Elasticsearch failures on delete
The Quotes module's save logic was not retrieving the beans before
attempting to flag them as deleted. How SugarBean::mark_deleted and
SugarBean::mark_undeleted work is questionable as it does seem to set
SugarBean::deleted property but not SugarBean::id.

Nevertheless it makes sense to perform a SugarBean::retrieve first as
this will imply any team security and ACL rules which seem to be
bypassed with the current logic.

Leaving mark_(un)deleted as is for now to minimize the impact of the
fix.