Opted out email addresses are valid (MAR-3943)
The client-side utility method that replicates SugarEmailAddress::getPrimaryAddress() was incorrectly suppressing opted out email addresses. To accurately replicate the behavior of SugarEmailAddress:getPrimaryAddress(), opted out email addresses may be returned.

When email invitations are sent, the server will use the email address returned by SugarEmailAddress::getPrimaryAddress() to get the email address of the invitee. We do not want to suggest to users in the UI that recipients with opted out email address will not receive the invitation when, in fact, they will. Opted out email addresses are only suppressed when sending campaign email; not notifications, invitations, or transactional email.