BR-4129: Token randomness and openssl non-randomness
This fix adds the random_compat library as a polyfil for php 7
random_bytes functionality. Because of a bug in PHP, openssl has been
scraped from random byte sources as a security measure which was the
primary issue on this ticket needed to be resolved.

Additionally a proper UUID library has been added to to replace create_guid
using the same random_compat polify. This allows us to replace the token
generation for OAuth2 but maintain the current format.

On a later stage the OAuth2 token format will drastically change adding
signartures, different length, ... However to minimize the impact of
this ticket we stick to the current UUID format paradigm of our tokens.