[BR-4394] FilterApi returns empty dataset in case of max_num=-1
Prior to the resolution of BR-3907, max_num=-1 was handled as a side effect of adding +1 to the max_num in Filter API in order to see whether there's more records in the data set, and then ignoring empty() value of the limit in SugarQuery.

The fix is to explicitly implement handling negative max_num on the API level.