Fixed a bug in Link2 (MAR-4045)
The bug described below only occurred when a `Link2` instance was used in the following sequence:
# `Link2::resetLoaded`
# `Link2::get`
# `Link2::getBeans` with parameters
# `Link2::getBeans` without parameters

We start with a link that hasn't been loaded. I chose to explicitly reset the loaded flag. Next we call
`Link2::get` to return the IDs of any linked records. This has the effect of changing
`Link2::loaded` to `true` while `Link2::$beans` is `null`. Next we call `Link2::getBeans` with
parameters. Before fixing this bug, this had the effect of changing `Link2::$beans` to an empty array. Next
we call `Link2::getBeans` without parameters. The `Link2` instance believed it had been loaded
because `Link2::beansAreLoaded` returns `true` when `Link2::$beans` is an array. We're simultaneously
using `Link2::$loaded` and `Link2::beansAreLoaded` to indicate if a link has been loaded, but they
aren't always in sync. So the current state of `Link2::$beans` was returned instead of executing the query
and populating `Link2::$beans` with its results.

After fixing this bug, calling `Link2::getBeans` with parameters will not change `Link2::$beans` to
an empty array. This allows `Link2::beansAreLoaded` to return `false` when we call `Link2::getBeans`
without parameters, so `Link2::$beans` will be populated with the data from `Link2::$rows` before it
is returned.