BR-4621: Fix uncontrollable growth of fts_queue table
During index connection timeouts, the records are send back the queue
from BulkHandler and the previous one's are cleaned up as the indexing
request itself was succesful from Indexer's point of view.

However the logic within Indexer was actively checking the connection
availability sending back the remainder of the records as well without
purging the previous ones causing a build up.

This logic will be re-addressed in another fix to actual change the
BulkHandler logic a bit more. The requeue functionality in BulkHandler
is required and cannot be removed at this point as this is needed for
inline requests when ES is truly unavailable (otherwise we will miss
record updates).