BR-4642: Disable refresh during reindexing
Introduce the ability to disable refresh_interval when a reindex is
triggered as this indicates a huge bulk insert. By default we use the -1
value to completely disable the refresh, however this can be overriden
in $sugar_config if desirable. Additionally index.refresh_interval can
be set using index_settings as well (this was already the case, but can
be a good tuning parameter for hosted environments).