BR-4673: Disable wildcard analyzers for potential big fields
This fix disables the ngram tokenization on field type text, longtext
and tinymce (html) fields. Once the tokenizations is optimized this
feature will be enabled again.

The MultiFieldHandler can always be overriden in an upgrade safe fasion
if this is still wanted by a customer. However they are on their own to
assess the risk regarding index throughput and index size.