Revert 7_7_2 ES Changes (#32661)
* Revert "Merge pull request #32476 from skymeyer/elastic772"

This reverts commit 6ce5b50ebc75e3688d4245e12697901e98fa8ad6, reversing
changes made to 406f3d4539c8e2722caa491883cc02fc8bef1254.

* BR-4622: (SI 77221) [Full Text Search] System permits Cron execution in parallel

* BR-4621: Fix elastic search connection availability

Quick fix to always assume the connection is available and not used the
cached current connection state which may be incorrectly influenced
because of timeouts.

The connection state is updated and checked for when actually forced
which happens only in background/cron processes. This fix avoids the ES
backend to be declared unavailable because of timeouts during long index
processes.

* BR-4621: Fix uncontrollable growth of fts_queue table

During index connection timeouts, the records are send back the queue
from BulkHandler and the previous one's are cleaned up as the indexing
request itself was succesful from Indexer's point of view.

However the logic within Indexer was actively checking the connection
availability sending back the remainder of the records as well without
purging the previous ones causing a build up.

This logic will be re-addressed in another fix to actual change the
BulkHandler logic a bit more. The requeue functionality in BulkHandler
is required and cannot be removed at this point as this is needed for
inline requests when ES is truly unavailable (otherwise we will miss
record updates).

* BR-4620: Administration commands for search

Adding two new commands to help the administration of globalsearch:
1. Ability to enable/disable a module for search
2. Queue cleanup command