Hack to avoid massupdatable fields to be added to the main context (SC-6206)
Massupdate is being initialized in the same context of the list view, thus
it will try to add its fields into the context when it shouldn’t, because
those fields belong to a different context/panel that has nothing to do
with.
Since this will require a major refactor on the mass update component,
this hack will prevent the fields to be added, while also preventing the
context from adding `view=true` in the request params.