Fix record view test to not fail due to async validation (SC-6218)
We need to mock this because the previous test was assuming the validation
of the model was always `false`, thus not triggering the save.
Although, the next test that was added for the locked fields feature, is
adding all the required fields to the model, thus making it valid.
Since `doValidate` method is asynchronous we need to make sure that tests
don’t continue without finishing the full flow.