Ignore links with no data (MAR-4502)
This is necessary because sidecar passes deltas for links without regard for whether or not an action is empty. `ModuleApi` will receive `$args['link_name'] = array('add' => array(), 'create' => array(), 'delete' => array())`. Any of the actions within a link argument may be empty or may contain data to process. If `ModuleApi` does not account for it, then a blank instance of the related bean will be returned from `BeanFactory::retrieveBean` in `RelateRecordApi::createRelatedLinks` -- because $id is null -- and an attempt is made to link the related record that isn't actually a record at all. By handling this early, in `ModuleApi`, we can avoid triggering exceptions that may originate from any number of paths, like custom module endpoints for linking records. In addition, there is a performance gain from avoiding all of the unnecessary function calls and memory allocation when it is clear that there is no operation to perform.