BR-4972: Use a separate session for BWC
By using a BWC session id which is not tied to the access token, there
will be no unnecessary calls to bwcLogin endpoint on every access token
change.

By setting the BWC session timeout to the refresh_token lifetime we
basically sync up the session life time for BWC with the OAuth2 lifetime
regardless of the session.cookie_lifetime settings which can cause
havoc.

This means that bwcLogin will only be called once within an OAuth2
refresh token life cycle. This will give more confidence that BWC session
values are stable instead of creating a new one on every access token
change which probably has led to other bugs as the past (i.e. import and
optimistic locking).

As a drawback there is a full separation of the BWC and OAuth2 session.
This is probably okay as there should be no session intersection between
BWC and sidecar modules in the first place.

The current fix lacks proper refresh token lifetime synchronisation. To
finetune this we probably should set the BWC cookie expiration time
based on the already elapsed time. This requires to be able to corelate
an access token to a refresh token to determine this value. Not sure if
that is present at this point.

There are some more quirks which need to be resolved: i.e. calling
logout end point requires a valid token. If an access token is expired
the logout does not properly trigger cleaning up existing sessions. This
was already the case before this change.