[BR-5260] Fixed the failure of Bug43466Test
The failure doesn't seem directly related to PHP 7.2 but happened if the failing test was preceeded by Bug53218Test. The latter would set the 'skip_index_rebuild' DB option to TRUE which would lead to skipping the index rebuild in the failing test.

The solution is to restore the option in repairDatabase.php after rebuilding the table.

Also the failing test was recovered since its output during the failure contained a lot of irrelevant details. Particularly, the vardefs copy/pasted from the Calls module.