BR-5394: Authenticated Arbitrary Read of File System Files (#34346)
* BR-5394: Authenticated Arbitrary Read of File System Files

When hitting module=Connectors&action=CallRest&url=/etc/passwd arbitrary
files can be read from the file system. This is already protected by
InputValidation, however the soft_fail flag needs to be turned off to
be properly enforced.

This commit adds additional coverage to ensure only valid URL's can be
passed in on the Symfony Url Constraint Validator.

* BR-5394: {old} unit test fixes to prep for disable soft_fail

Address {old} unit tests to be able to pass when soft_fail is disabled
on the InputValidation framework.

Add the ability to flush the ConstraintValidatorFactory cache. This is
necessary for unit tests to be able to reinitialize proper constraint
instantiation. During regular code execution for performance reasons all
Constraint objects are cached for reusage.