BR-5419: Add input_validation log channel
This change will allow an admin to control the log level for
InputValidation regardless of the global log level. Changing the log
level is done by using the following config_override directive:

$sugar_config['logger']['channels']['input_validation']['level'] = 'warning';

This example allows to log all violations which are emitted on WARNING
level when soft_fail is enabled, even if the global log level is set to
fatal only.