MAR-4664: UserSignatures extends Data.Bean to provide custom save method
UserSignatures provides a custom save model which will make a ping request if the is_default attribute has changed in the model.  This is necessary as the saving of the UserSignatures with a new default value will change the user's preferences making the metadata out of sync.  The ping request is to update the user pref hash in memory after the signature save and before making a fetch request. If a view knows it just performed an action that requires updating the metadata or user prefs in some way, it is convention to make a ping on success to prevent the next action from triggering a 412.