[BR-5491] [DB2] Fatal error when switching opportunity mode
Remove all references to the connection resource when disconnecting.

The issue is reproducible because `MetaDataManager::refreshCache()` is executed in the shutdown phase after `sugar_cleanup()` has already disconnected from the database. It establishes a new connection while the previous connection resource is not completely freed. It causes some issues inside the DB2 driver.