BR-5496: Add sugar version in the header of HTTP requests to Elastic
This PR will add an HTTP User-Agent header in the form of
SugarCRM/x.x.x.x where x.x.x.x is the version number of the sugar
instance. This can be used for identification purposes to orchestrate
any (reverse) proxies which sits in front of the ES cluster we are
connecting to.