BR-5366: Prevent DoS and apply proper platform validation (#34398)
* BR-5366: Add platform constraint validator

* BR-5366: Apply $api->platform validation

The main symptom of this ticket is caused by not validating the platform
value properly resulting in an HTTP 500 when trying to create a cache
file with an very long name which will fail.

A previous ticket BR-3435 already introduced the functionality to
enforce validating the requested platform against the known list. This
only happens on the OAuth2Api, but not on all the other API of which the
MetaDataApi is the issue of this specific ticket.

Because we never enforced validation of the platform before and some
customizations may rely on this "feature", the above check needs to be
explicitly enabled using $sugar_config['disable_unknown_platforms'] =
true.

This ticket adds a proper validation method on ServiceBase and enforce
the validation when the $platform is set. The platform constraint
validator honors the above mentioned $sugar_config flag. If this flag is
not enabled we still perform generic validation to ensure only valid
characters and length is enforced to fit in our db column.

* BR-5366: Re-address platform validation as per BR-3435

Apply the newly created ServiceBase::validatePlatform call replacing the
earlier implementation on the OAuthApi end point. This change is fully
backwards compatible but now relies on the Validator service directly
which was not available earlier.