Created a list-header template for RowactionsCreateField (SC-6429)
Since no list-header template was found, render was falling back to a template that rendered the fields found in the fieldset. RowactionsCreateField overrides the standard fieldset behavior to push its buttons as fields. When the fieldset was rendered, its buttons were rendering, even in the list header column. The new list-header template is a straight copy of the list-header template from RowactionsField, which RowactionsCreateField extends. RowactionsCreateField now behaves the same as RowactionsField in the list header.