BR-5565: Remove deprecated usage of include_in_all
For Elastic 5.x the usage of include_in_all is deprecated. By using this
statement in the mapping deprecation warning messages are shown in the
Elasticsearch log files.

Because we do not rely on the _all field (which is explicitly disabled
in the mapping for each module), explicitly setting include_in_all to
false is redundant. We can safely remove these statements while keeping
_all to disabled as well (which is also deprecated).