BR-5537: Support same fields with different mapping
This commit minimized the risk of field mapping collisions. We are using
the base fields (ie $bean->name) with the same mapping across the board
and use the copyTo functionality to implement per module/(multi)field
mapping. This will allow more flexibility for the future.

In some situations collisions are unavoidable. For example the email
field requires an "object" type which cannot be hidden in a multi
field. The only solution for unavoidable collisions is using separate
indices to isolate the colliding modules. For the core modules this
should never happen.