BR-5562: Removal of ElasticaFactory
BR-5147 introduced a ElasticaFactory for Elasticsearch 5.x support. This
commit removes this factory to have direct references to \Elastica
objects instead. ElasticaFactory has been a reminent from the attempt of
supporting ES 1.x and ES 5.x at the same time. However this is
architecturally impossible and therefor we remove ElasticaFactory as it
adds too much complexity and is not really suited for custom code usage
in the first place.

This commit also contains some namespace cleanups and removing the usage
of the deprecated QueryBuilder::FIELD_SEP and ::PREFIX_SEP.