BR-5562: Fix not_analyzed multi field handler
As the base of a multi field is no longer indexed, the fields which
require a not_analyzed value need to be explicitly defined in the
mapping now. This commit removes the special not_analyzed handling from
MultiFieldHandler and treats (gs_)not_analyzed fields as a regular
mapping type now. This also implies that no explicit fields need to be
added anymore to the highlighter section and we can just rely on
wildcarded *.gs_not_analyzed field declaration directly.

We also cleanup the isLongFieldTypes and isFieldSearchable as they are
no longer used.