BR-5537: Support same fields with different mapping
Address usage of object field properties which can still collide with
other field names using different mappings. For such object properties
we explicitly prefix them without the usage of a primary field like we
do for multi fields using copy_to.

This PR adds handling for _source name remaps and field normalization to
be able to properly retrieve the data. Therefore a dedicated
ResultParser got introduced which can be injected.

This ResultParser handles both _source and _highlight. The logic for
parsing has been moved out of the Highlighter itself. Some additional
cleanups have been done on the Highlighter as the usage of an Abstract
class was kind of dubious.

Additionally SearchFields has been updated to use new SearchField
objects so it is easier to process this list on mulitple levels,
especially on the Query part. This part was redone is it relied on the
"field normalization" logic formerly present in Highlighther which has
moved into the ResultParser where it actually belongs.

The logic in MultiMatchQuery regarding Search field handling was an
afterthought process. By introducing SearchField objects and
SearchFields collection object, there is not need to reprocess the plain
searchFields array anymore to try and determine the requested bean
fields.

This commit also contains the required unit test fixes for the touched
areas and cleans up some field references for KBContents specific
queries from the legacy code.