BR-5598: Address Unit Test Failures
Although this is not directly caused by the tests themselves, the array
to string conversion as reported by the stack trace is triggered when
error handling index error exceptions from the Elasticsearch backend.

This index error exceptions are related to something within the
Elasticsearch cluster itself which may be unstable during test
execution. Either way by addressing the array->string conversion we will
get the proper output in FATAL logs if this should happen again.

During investigation an additional fix for non-existing array index was
found in the "bin/sugarcrm search:status" CLI command which has been
addressed which is triggered when ES is not available.