[BR-5686] [PHPUnit] Fixed failures in ForecastsChartApiTest
The test doesn't set up proper fixtures. It works for normalized schema but dosn't work for the new one. The `reports_to` field of the usre is populated with the pre-gererated value but the user with this ID doesn't exist.

We need to skip adding manager to the team if the manager doesn't exist.