[BR-5686] [PHPUnit] Fixed failures in TeamBasedACLVisibilityTest
The test set up the visibility mock prior to changing current user. The normalized visibility implementation uses `global $current_user` directly, so it didn't cause the failure. The denormalized implementation captures the reference to the current user and uses it. Therefore, the mock should be created after setting the user (via callback).