MAR-4874: Synchronizes the email's sender and recipients on retrieve
From 7.0.x through 7.9.x, REST API clients, like OPI/LPI, could create emails using Emails API by specifying the sender and recipients with from_addr_name, to_addrs_names, cc_addrs_names, and bcc_addrs_names in the request body. This led to the sender and recipients being stored in the emails_text table but not synchronized to the emails_email_addr_rel table. To correct this, the sender and recipients are synchronized to the emails_email_addr_rel table when the email is retrieved. This correction only happens when the emails_email_addr_rel table lacks data that is found in the emails_text table. And it only occurs once per email.