[BR-5810] Filters.Teamset REST API test fails on DB2 (#35312)
The test assumes that the primary team is always the first on the list, however the teams are ordered alphabetically.

Prior to fixing BR-5757, team names were sorted in PHP where the NULL value of the team name was cast to empty string and therefore was always the first. As of BR-5757, team names are sorted in SQL, and DB2 puts NULLs at the end of the list.