[BR-5762] (SI 79828) TeamSecurity applied to WHERE clause instead of SELECT clause causes team security to be applied incorrectly when querying based on a relationship (#35127)
* [BR-5762] (SI 79828) TeamSecurity applied to WHERE clause instead of SELECT clause causes team security to be applied incorrectly when querying based on a relationship

The code like `$query->where()->addRaw($sql)` sets the raw WHERE condition on the top level of SugarQuery. If called multiple times (e.g. on primary and related tables in a query), only the last condition remains effective.

Prior to #30032, `TeamSecurity` used `$query->whereRaw($sql)` which creates a `new` raw expression in the top level container, not sets a raw value in it.

The `SugarQuery_Builder_Where` class can act both as condition (raw and non-raw) and container for other condition. In order to prevent issues like this from reproducing, when adding a condition to the container or setting raw expression, we need to make sure we're not losing previously assigned conditions or raw expression.

Eventually, it needs to be reworked so that raw condition, non-raw condition (e.g. `equals()`) and container (OR and AND) are represented by separate classes, and the raw class is immutable.

* [BR-5686] Added more test cases, renamed fixtures

The `SugarQuery` test covers the cases when the primary record is visible to the user but the related one is not and the other way around.

The changes in generated fixtures do not affect the Reports module because (let's call it a known issue) reports always use INNER JOIN for related tables, so if the related record is invisible, the primary one will be not visible as well.

* [BR-5762] Added a test case