PX-11: Make merge view specify the list of fields to request
merge-duplicate's getFieldNames() method is reliant on the mergeFields
property, which is populated in _initializeMergeFields(). Move the call
to this method above the parent initialize(), which calls getFieldsNames()
with an empty list of mergeFields.

Without this change, Merge Duplicates sends an empty list of fields
to the Filter API, causing issues like BR-5831.

Signed-off-by: Bob Wombat Hogg <rhogg@sugarcrm.com>