[MAR-5001] Don't clone the related model until it is time to use it
The model was being cloned when the EmailField instance was initialized, which was happening before the model was completely hydrated by the server response. The fields that were not yet populated were never being passed to the related model, which was the root cause of the bug in MAR-5001. It may not be necessary to clone the model at all, but we're continuing to do it to reduce regression risk.