[BR-6098] Use DBManager instead of DBAL Schema Manager for schema introspection
In order to introspect the schema on Oracle or IBM DB2, any of the implementations above will need to have an explicitly specified user or schema name respectively which are taken from the connection configuration. Unlike DBManager, DBAL doesn't have access to the configuration since it reuses the connection resource established by DBManager.

TL;DR, we cannot use DBAL Schema Manager for now due to the fact that DBAL is our secondary DB layer.