[MAR-5060] converted_opportunity_name was too long of a field name
Oracle doesn't support table names, field names, or aliases longer than 30 characters. Sometimes we create aliases with suffixes like _owner, _mod, etc. converted_opportunity_name_owner is 32 characters, so it broken tests on the Oracle stack. The name of the field was shortened so that it would never result in an alias longer than 30 characters. In the long term, all aliases should be generated using DBManager::getValidDBName() so that we have a little more flexibility with our field name lengths.