[BR-6406] Fixed test failures on Oracle
Prior to sebastianbergmann/comparator#58, PHPUnit considered string like "2" and "2.000000" equal. The OCI8 driver, on the other hand, unlike other DB drivers truncates trailing zeroes in numbers.

After the upgrade of PHPUnit, some tests started failing. The tests are fixed to not expect numbers to be formatted in any specific ways since it always worked like this on Oracle.