[BR-6601] Removed SugarBeanTest::testLoadAutoIncrementValuesWithoutAutoincrement()
The test never worked since a private method cannot and couldn't be mocked. It tested that a method wouldn't be called under certain circumstances which is an implementation detail. At the same time, even if it was called, it wouldn't produce any side effects because it has the same check as being tested in the calling method.