[BR-6613] Disable audit on the SchedulersJobs and Notifications modules
The audit of SchedulersJobs was enabled since its creation in 7.7.0 (https://github.com/sugarcrm/Mango/pull/24353) but it didn't cause any side effects since the audit fields on a job record (fallible and assigned_user_id) aren't usually changed. As of BR-5877, the values of the audited fields are logged upon the record creation as well which results in extensive growth of the audit tables.

The same reasoning applies to Notifications which an internal entity which doesn't mutate and isn't subject for audit in the first place.