[MAR-5173] Format saved filters for API clients and unformat filters before saving
Filters for the email participants fields must include the names and email addresses of participants. These values cannot be persisted with the filter definition or it would violate GRPC. Therefore, the values must be removed from the filter definition when saving to the database and dynamically loaded when sending the filter definition to the client.

ApiHelper must use `getModulename` in order to find FiltersApiHelper because the Filters module has a `module_name` field which conflicts with the field of the same name that is inherited from `SugarBean`.