CS-176: Refactor DefaultDashboardInstaller
This commit *has* to go into the product before
writing an upgrade script that uses
DefaultDashboardInstaller::buildDashboardFromFile().
Because on a pro-to-ent conversion, the version of
that file from current 9.1.0 is used, not the new
one.

This won't be a problem when the real 9.1.0 goes
out, because by then this commit will have been
merged and already present in Pro. But for now
you have to test this by building from this
commit, THEN running the upgrader against
the subsequent commit that adds the upgrade
script.

Sometimes, technology makes life easy.
Sometimes, it makes life harder.

Signed-off-by: Bob Wombat Hogg <rhogg@sugarcrm.com>