CS-338: Update Mean time to resolution in cases Added 2 new fields: - Hours to Resolution - Business Hours to Resolution
Both fields are of type decimal (so 30 min will be 0.5 hrs).
Both fields should be visible, but readonly, on Record View.
Both fields should be available on list views, but not on the list by default.
Both fields should only be for serve.
If a case is reopened, and closed again, then re-calculate both fields.

Let's keep the existing field 'Time to Resolution' from 9.1 as is. We'll open a separate ticket to remove it.